Still single-log-output instead of both console+file
